test(security): isolate npm --ignore-scripts test to avoid child_process mock pollution
The module-wide vi.mock('child_process') broke 4 pre-existing subprocess
tests in versions.test.ts under the node runner. Move the mock + test into
versions.installscripts.test.ts so it can't leak.
